backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1